Rename: svg 이미지 폴더 위치 변경
