+++ qrumble/src/router.jsx
@@ -1,39 +1,39 @@
-import { createBrowserRouter, Navigate } from 'react-router-dom';
-import Login from './login/pages/Login';
-import HomePage from './home/HomePage';
-import Splash from './Splash';
-import FindId from './login/pages/FindId';
-import ResetPassword from './login/pages/ResetPassword';
-import SignUp from './login/pages/SignUp';
-import NewPassword from './login/pages/NewPassword';
+import { createBrowserRouter, Navigate } from "react-router-dom";
+import Login from "./login/pages/Login";
+import HomePage from "./home/pages/HomePage";
+import Splash from "./Splash";
+import FindId from "./login/pages/FindId";
+import ResetPassword from "./login/pages/ResetPassword";
+import SignUp from "./login/pages/SignUp";
+import NewPassword from "./login/pages/NewPassword";
 
 const router = createBrowserRouter([
   {
-    path: '/',
+    path: "/",
     element: <Splash />,
   },
   {
-    path: '/login',
+    path: "/login",
     element: <Login />,
   },
   {
-    path: '/find-id',
+    path: "/find-id",
     element: <FindId />,
   },
   {
-    path: '/reset-password',
+    path: "/reset-password",
     element: <ResetPassword />,
   },
   {
-    path: '/reset-password/new',
+    path: "/reset-password/new",
     element: <NewPassword />,
   },
   {
-    path: '/sign-up',
+    path: "/sign-up",
     element: <SignUp />,
   },
   {
-    path: '/home',
+    path: "/home",
     element: <HomePage />,
   },
 ]);
